feat: 출석 랭킹 동시성 문제 해결 - redis 사용
issue: #453

+++ src/main/java/keeper/project/homepage/about/dto/AttendanceResultDto.java
@@ -20,7 +20,7 @@ public class AttendanceResultDto {
   private String thumbnailPath;
   private String greetings;
   private Integer continuousDay;
-  private Integer rank;
+  private Long rank;
   private Integer point;
   private Integer rankPoint;
   private Integer continuousPoint;
